Refactor: use barrel imports for utility functions
