+++ lib/LocalAuthModule.js
@@ -1,11 +1,9 @@
 import _ from 'lodash'
 import { AbstractAuthModule } from 'adapt-authoring-auth'
 import apidefs from './apidefs.js'
-import { compare } from './utils/compare.js'
+import { compare, getRandomHex, validate } from './utils.js'
 import { formatDistanceToNowStrict as toNow } from 'date-fns'
-import { getRandomHex } from './utils/getRandomHex.js'
 import PasswordUtils from './PasswordUtils.js'
-import { validate } from './utils/validate.js'
 /**
  * Module which implements username/password (local) authentication
  * @memberof localauth
