+++ lib/PasswordUtils.js
@@ -1,9 +1,7 @@
 import { App } from 'adapt-authoring-core'
 import bcrypt from 'bcryptjs'
-import { compare } from './utils/compare.js'
-import { getRandomHex } from './utils/getRandomHex.js'
+import { compare, getRandomHex, validate } from './utils.js'
 import { promisify } from 'util'
-import { validate } from './utils/validate.js'
 
 /** @ignore */ const passwordResetsCollectionName = 'passwordresets'
 /**
